refactor: separate stacked-area-status-chart
