+++ src/app/units/states/analytics/directives/charts/stacked-area-status-chart/stacked-area-status-chart.component.html
@@ -1,16 +1,33 @@
-<ng-charts-area-chart-stacked
-  class="reverse-legend-order"
-  [animations]="true"
-  [gradient]="false"
-  [legend]="true"
-  [results]="weeklyData"
-  [scheme]="$any(colorScheme)"
-  [showXAxisLabel]="true"
-  [showYAxisLabel]="true"
-  [tooltipDisabled]="false"
-  [xAxis]="true"
-  [xAxisLabel]="XAxisLabel"
-  [yAxis]="true"
-  [yAxisLabel]="YAxisLabel"
->
-</ng-charts-area-chart-stacked>
+<mat-card-header class="text-pretty xl:max-w-[calc(50%)]">
+  <mat-card-title>Task Completion Over Time</mat-card-title>
+  <mat-card-subtitle class="mb-2.5">
+    <p>
+      View the task completion statuses across all tasks over time, filtered by campus and/or
+      tutorial.
+    </p>
+  </mat-card-subtitle>
+</mat-card-header>
+@if (snapshots.length) {
+  <div class="min-h-100">
+    <ng-charts-area-chart-stacked
+      class="reverse-legend-order"
+      [animations]="true"
+      [gradient]="false"
+      [legend]="true"
+      [results]="weeklyData"
+      [scheme]="$any(colorScheme)"
+      [showXAxisLabel]="true"
+      [showYAxisLabel]="true"
+      [tooltipDisabled]="false"
+      [xAxis]="true"
+      [xAxisLabel]="XAxisLabel"
+      [yAxis]="true"
+      [yAxisLabel]="YAxisLabel"
+    >
+    </ng-charts-area-chart-stacked>
+  </div>
+} @else {
+  <div class="flex h-full items-center justify-center">
+    <mat-progress-spinner mode="indeterminate" [diameter]="48"></mat-progress-spinner>
+  </div>
+}
